+++ src/MiscScroll.sol
@@ -8,4 +8,7 @@ library MiscScroll {
 
   // https://scrollscan.com/address/0x782559e349b084bB7C07c08404aE6E3436cDAE2E
   address internal constant PROXY_ADMIN = 0x782559e349b084bB7C07c08404aE6E3436cDAE2E;
+
+  // https://scrollscan.com/address/0x63B20270b695E44Ac94Ad7592D5f81E3575b93e7
+  address internal constant PROTOCOL_GUARDIAN = 0x63B20270b695E44Ac94Ad7592D5f81E3575b93e7;
 }