+++ src/ts/MiscArbitrum.ts
@@ -13,3 +13,5 @@ export const CREATE_3_FACTORY = '0x3b56998Ec06477704622ca8e2eA1b4db134cec32';
 
 // https://arbiscan.io/address/0xbbd9f90699c1FA0D7A65870D241DD1f1217c96Eb
 export const PROTOCOL_GUARDIAN = '0xbbd9f90699c1FA0D7A65870D241DD1f1217c96Eb';
+
+export const CHAIN_ID = 42161;